requirements.txt
paho-mqtt==1.3.1
python-dotenv==0.8.2
